Move debug code to a separate file
